[refactor](be) Delete old ES scan path and consolidate ES code under format/table/es/
### What problem does this PR solve?

Problem Summary: After unifying ES scans through FILE_SCAN_NODE, the old
ES_HTTP_SCAN_NODE scan path (EsScanner, EsScanOperatorX, pipeline
registration) became dead code. Additionally, ES-related reading code was
scattered across be/src/exec/es/ and be/src/format/table/.

Changes:
- Deleted old scan path: EsScanner, EsScanOperatorX (4 files, ~520 lines)
- Removed ES_SCAN_NODE/ES_HTTP_SCAN_NODE case from pipeline_fragment_context
- Removed EsScanLocalState template instantiations and operator declarations
- Moved all ES reading code (ESScanReader, ScrollParser, ESScrollQueryBuilder,
  EsHttpReader) from be/src/exec/es/ and be/src/format/table/ to
  be/src/format/table/es/
- Updated all include paths accordingly

### Release note

None

### Check List (For Author)

- Test: BE compilation verified, all 7 ES regression tests pass
- Behavior changed: No
- Does this need documentation: No
